New useCaravaggioIfAvailable hook
